Quote variables in conan_cmake_detect_vs_runtime
Account for the possibility that variables are empty, or have values which are themselves variable names, in conan_cmake_detect_vs_runtime.

There are two issues this quoting resolves. One is errors resulting from empty variables - for exampe, if `CMAKE_CXX_FLAGS` is empty, the `string` call would have failed because it would have appeared to have too few arguments - quoting the final argument fixes this.

The other is that in the context of `if`, a lot of operators try to resolve their argument as a variable name first (but only if it's unquoted), meaning that `if(${build_type} STREQUAL "DEBUG")` would have unexpected results if `build_type` is set to `DEBUG` and there is a variable `DEBUG` set to `YES`, say. Quoting the variable expansion prevents this behaviour.