Limit compile options considered for libcxx detection
The COMPILE_OPTIONS directory property can contain generator
expressions, which are unlikely to be permitted by the compiler as valid
options. To limit the cases where this is an issue, only pull out the
options that might have an impact in conan_cmake_detect_unix_libcxx.

Note: we actually pull out -std= options as well as -stdlib options,
which may not be necessary, but is unlikely to be an issue.