Pass through unknown architectures
Conan supports non-standard architectures through the installation of
custom settings.yml files. However, conan-cmake currently expects to be
able to translate the architecture from the CMake naming scheme to the
Conan naming scheme, and fails with an error if the architecture is
unknown.

This patch resolves the issue by making the script simply pass through
the CMake architecture to Conan unchanged in these cases.