conan-provider.cmake: introduce a new policy scope, set min cmake version
cmake has the notion of "policy scopes" to limit the scope of the policies.
this patch introduces a module-level policy scope and sets the minimum
required cmake version to 3.24, so the cmake features depend on policy settings
such as `if(... IN_LIST ...)` behaves as expected even if the consuming project
does not declare a cmake minimum version requirement. also, this change now
enforces the version requirement so the code would fail immediately if the
cmake version is less than the minimum.

Signed-off-by: Mustafa Kemal Gilor <hello@mkg.dev>