Add support for CONAN_INSTALL_BUILD_CONFIGURATIONS
Let the user or project override the set of build types we execute the
conan install command for. This is most useful for enabling projects to
build only one configuration when using a multi-config generator, but it
can also be used to install for multiple build types with a single-config
generator.